adding custom exceptions and responses

+++ MyLeaveManagement.Application/Features/LeaveAllocation/Handlers/Command/DeleteLeaveAllocationCommandHandler.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
 using MediatR;
+using MyLeaveManagement.Application.Exceptions;
 using MyLeaveManagement.Application.Features.LeaveAllocation.Requests.Command;
 using MyLeaveManagement.Application.Persistence.Contracts;
 
@@ -23,6 +24,8 @@ namespace MyLeaveManagement.Application.Features.LeaveAllocation.Handlers.Comman
         public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+            if (leaveAllocation == null)
+                throw new NotFoundException(nameof(Domain.LeaveAllocation), request.Id);
             await _leaveAllocationRepository.Delete(leaveAllocation.Id);
             return Unit.Value;
         }
