+++ MyLeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -5,8 +5,10 @@ using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
 using MediatR;
+using MyLeaveManagement.Application.Exceptions;
 using MyLeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using MyLeaveManagement.Application.Persistence.Contracts;
+using MyLeaveManagement.Domain;
 
 namespace MyLeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
 {
@@ -23,6 +25,8 @@ namespace MyLeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.Get(request.Id);
+            if (leaveType == null)
+                throw new NotFoundException(nameof(LeaveType), request.Id);
             await _leaveTypeRepository.Delete(leaveType.Id);
             return Unit.Value;
         }
